Update IIR_Butterworth_Main.cs
Changes have been made to allow the user to filter the signal. Please, keep in mind that pre and post filter data along with the time domain are now saved in 3 different txt files, so the results can be imported and used in other softwares, such as Matlab. The user needs to change the directory where the txt files will be saved. That is line 39: 

string dir = @"C:\Users\press\OneDrive\Desktop\IIR_Butterworth_C_Sharp";